Revert "add table eval and predict script" (#3062)
