--- ppocr/utils/dict/table_dict.txt
@@ -1,277 +1,0 @@
-←
-</overline>
-☆
-─
-α
-﻿
- 
-⋅
-$
-ω
-ψ
-χ
-(
-υ
-≥
-σ
-,
-ρ
-ε
-0
-■
-4
-8
-✗
-b
-<
-✓
-Ψ
-Ω
-€
-D
-3
-Π
-H
-║
-</strike>
-L
-Φ
-Χ
-θ
-P
-κ
-λ
-μ
-T
-ξ
-X
-β
-γ
-δ
-\
-ζ
-η
-`
-d
-<strike>
-h
-f
-l
-Θ
-p
-√
-t
-</sub>
-x
-Β
-Γ
-Δ
-|
-ǂ
-ɛ
-j
-̧
-➢
-⁡
-̌
-′
-«
-△
-▲
-#
-</b>
-'
-Ι
-+
-¶
-/
-▼
-⇑
-□
-·
-7
-▪
-;
-?
-➔
-∩
-C
-÷
-G
-⇒
-K
-<sup>
-O
-S
-С
-W
-Α
-[
-○
-_
-●
-‡
-c
-z
-g
-<i>
-o
-<sub>
-〈
-〉
-s
-⩽
-w
-φ
-ʹ
-{
-»
-∣
-̆
-e
-ˆ
-∈
-τ
-◆
-ι
-∅
-∆
-∙
-∘
-Ø
-ß
-✔
-∞
-∑
-−
-×
-◊
-∗
-∖
-˃
-˂
-∫
-"
-i
-&
-π
-↔
-*
-∥
-æ
-∧
-.
-⁄
-ø
-Q
-∼
-6
-⁎
-:
-★
->
-a
-B
-≈
-F
-J
-̄
-N
-♯
-R
-V
-<overline>
-―
-Z
-♣
-^
-¤
-¥
-§
-<underline>
-¢
-£
-≦
-­
-≤
-‖
-Λ
-©
-n
-↓
-→
-↑
-r
-°
-±
-v
-<b>
-♂
-k
-♀
-~
-ᅟ
-̇
-@
-”
-♦
-ł
-®
-⊕
-„
-!
-</sup>
-%
-⇓
-)
--
-1
-5
-9
-=
-А
-A
-‰
-⋆
-Σ
-E
-◦
-I
-※
-M
-m
-̨
-⩾
-†
-</i>
-•
-U
-Y
- 
-]
-̸
-2
-‐
-–
-‒
-̂
-—
-̀
-́
-’
-‘
-⋮
-⋯
-̊
-“
-̈
-≧
-q
-u
-ı
-y
-</underline>
-​
-̃
-}
-ν
