--- ppstructure/MANIFEST.in
@@ -1,9 +1,0 @@
-include LICENSE
-include README.md
-
-recursive-include ppocr/utils *.txt utility.py logging.py network.py
-recursive-include ppocr/data/ *.py
-recursive-include ppocr/postprocess *.py
-recursive-include tools/infer *.py
-recursive-include ppstructure *.py
-
