+++ tools/infer/predict_det.py
@@ -43,7 +43,7 @@ class TextDetector(object):
         pre_process_list = [{
             'DetResizeForTest': {
                 'limit_side_len': args.det_limit_side_len,
-                'limit_type': args.det_limit_type,
+                'limit_type': args.det_limit_type
             }
         }, {
             'NormalizeImage': {
